Reclassify 'unit' tests as 'vscode-integration' tests
* salesforcedx-vscode-lwc package
* salesforcedx-vscode-visualforce package
