+++ packages/salesforcedx-vscode-lwc/test/telemetry/telemetry.test.ts
@@ -6,14 +6,14 @@
  */
 import { assert, match, SinonStub, stub } from 'sinon';
 import TelemetryReporter from 'vscode-extension-telemetry';
-import { TelemetryService } from '../../../src/telemetry/telemetry';
+import { TelemetryService } from '../../src/telemetry/telemetry';
 
 describe('Telemetry', () => {
   let reporter: TelemetryReporter;
   let sendEvent: SinonStub;
 
   beforeEach(() => {
-    reporter = new TelemetryReporter('salesforcedx-vscode', 'v1', 'test567890');
+    reporter = new TelemetryReporter('salesforcedx-vscode', 'v1', 'test345390');
     sendEvent = stub(reporter, 'sendTelemetryEvent');
   });
 
@@ -26,7 +26,7 @@ describe('Telemetry', () => {
     const telemetryService = TelemetryService.getInstance();
     telemetryService.initializeService(reporter, true);
 
-    telemetryService.sendExtensionActivationEvent([1, 300]);
+    telemetryService.sendExtensionActivationEvent([0, 678]);
     assert.calledOnce(sendEvent);
   });
 
@@ -34,19 +34,33 @@ describe('Telemetry', () => {
     const telemetryService = TelemetryService.getInstance();
     telemetryService.initializeService(reporter, false);
 
-    telemetryService.sendExtensionDeactivationEvent();
+    telemetryService.sendCommandEvent('force_lightning_lwc_component_create');
     assert.notCalled(sendEvent);
+  });
+
+  it('Should send correct data format on sendCommandEvent', async () => {
+    const telemetryService = TelemetryService.getInstance();
+    telemetryService.initializeService(reporter, true);
+
+    telemetryService.sendCommandEvent('force_lightning_lwc_component_create');
+    assert.calledOnce(sendEvent);
+
+    const expectedData = {
+      extensionName: 'salesforcedx-vscode-lwc',
+      commandName: 'force_lightning_lwc_component_create'
+    };
+    assert.calledWith(sendEvent, 'commandExecution', expectedData);
   });
 
   it('Should send correct data format on sendExtensionActivationEvent', async () => {
     const telemetryService = TelemetryService.getInstance();
     telemetryService.initializeService(reporter, true);
 
-    telemetryService.sendExtensionActivationEvent([1, 300]);
+    telemetryService.sendExtensionActivationEvent([0, 678]);
     assert.calledOnce(sendEvent);
 
     const expectedData = {
-      extensionName: 'salesforcedx-vscode-visualforce',
+      extensionName: 'salesforcedx-vscode-lwc',
       startupTime: match.string
     };
     assert.calledWith(sendEvent, 'activationEvent', match(expectedData));
@@ -60,7 +74,7 @@ describe('Telemetry', () => {
     assert.calledOnce(sendEvent);
 
     const expectedData = {
-      extensionName: 'salesforcedx-vscode-visualforce'
+      extensionName: 'salesforcedx-vscode-lwc'
     };
     assert.calledWith(sendEvent, 'deactivationEvent', expectedData);
   });
