+++ packages/salesforcedx-vscode-visualforce/package.json
@@ -57,7 +57,8 @@
     "test": "node ../../scripts/run-test-with-top-level-extensions",
     "test:insiders":
       "cross-env CODE_VERSION=insiders node ../../scripts/run-test-with-top-level-extensions",
-    "test:unit": "node ../../scripts/run-test-with-top-level-extensions"
+    "test:vscode-integration":
+      "node ../../scripts/run-test-with-top-level-extensions"
   },
   "activationEvents": ["workspaceContains:sfdx-project.json"],
   "main": "./out/src/extension",
